extract addHeader to separate module
