+++ src/__tests__/gatsby-node.spec.js
@@ -2,8 +2,6 @@ const { KontentTestHttpService }
   = require('@kentico/kontent-test-http-service-js');
 
 const { sourceNodes } = require('../gatsby-node');
-const { customTrackingHeader } = require('../config');
-const { name, version } = require('../../package.json');
 const complexContentItemsFirstLanguageFakeReponse =
   require('./complexContentItemsFirstLanguageFakeReponse.json');
 const complexContentItemsSecondtLanguageFakeReponse =
@@ -12,21 +10,13 @@ const complexTypesFakeResponse =
   require('./complexTypesFakeResponse.json');
 const fakeTaxonomiesResponse = require('./fakeTaxonomiesResponse.json');
 
-describe('customTrackingHeader', () => {
-  it('has correct name', () => {
-    expect(customTrackingHeader.header).toEqual('X-KC-SOURCE');
-  });
-
-  it('has correct value according to package name and package version', () => {
-    const expectedHeaderValue = `${name};${version}`;
-    expect(customTrackingHeader.value).toEqual(expectedHeaderValue);
-  });
-});
 
 describe('sourceNodes', () => {
   const dummyCreateNodeID = jest.fn();
   dummyCreateNodeID.mockReturnValue('dummyId');
-  const schemaDummy = { buildObjectType: jest.fn((input) => ({ data: input }))};
+  const schemaDummy = {
+    buildObjectType: jest.fn((input) => ({ data: input })),
+  };
 
   const dummyCreation = {
     actions: {
@@ -36,126 +26,6 @@ describe('sourceNodes', () => {
     createNodeId: dummyCreateNodeID,
     schema: schemaDummy,
   };
-
-  describe('tracking header tests', () => {
-    const fakeEmptyResponseConfig = new Map();
-    fakeEmptyResponseConfig.set(
-      /https:\/\/deliver.kontent.ai\/.*\/items/,
-      {
-        fakeResponseJson: {
-          items: [],
-          modular_content: {},
-          pagination: {
-            continuation_token: null,
-            next_page: null,
-          },
-        },
-        throwError: false,
-      });
-    fakeEmptyResponseConfig.set(
-      /https:\/\/deliver.kontent.ai\/.*\/types/,
-      {
-        fakeResponseJson: {
-          types: [],
-          pagination: {
-            continuation_token: null,
-            next_page: null,
-          },
-        },
-        throwError: false,
-      });
-    fakeEmptyResponseConfig.set(
-      /https:\/\/deliver.kontent.ai\/.*\/taxonomies/,
-      {
-        fakeResponseJson: {
-          taxonomies: [],
-          pagination: {
-            continuation_token: null,
-            next_page: null,
-          },
-        },
-        throwError: false,
-      });
-
-    const fakeEmptyTestService =
-      new KontentTestHttpService(fakeEmptyResponseConfig);
-
-
-    it('does add tracking header', async () => {
-      const deliveryClientConfig = {
-        projectId: 'dummyEmptyProject',
-        httpService: fakeEmptyTestService,
-      };
-
-      await sourceNodes(
-        dummyCreation,
-        {
-          deliveryClientConfig,
-          languageCodenames: ['default'],
-        },
-      );
-
-      expect(deliveryClientConfig.globalQueryConfig.customHeaders)
-        .toContainEqual(customTrackingHeader);
-    });
-
-    it('does update tracking header value', async () => {
-      const deliveryClientConfig = {
-        projectId: 'dummyEmptyProject',
-        httpService: fakeEmptyTestService,
-        globalQueryConfig: {
-          customHeaders: [{
-            header: customTrackingHeader.header,
-            value: 'dummyValue',
-          }],
-        },
-      };
-
-      await sourceNodes(
-        dummyCreation,
-        {
-          deliveryClientConfig,
-          languageCodenames: ['default'],
-        },
-      );
-
-      expect(deliveryClientConfig.globalQueryConfig.customHeaders)
-        .toContainEqual(customTrackingHeader);
-      expect(deliveryClientConfig.globalQueryConfig.customHeaders.length)
-        .toEqual(1);
-    });
-
-    it('does not influence other tracking header value', async () => {
-      const anotherHeader = {
-        header: 'another-header-name',
-        value: 'dummyValue',
-      };
-      const deliveryClientConfig = {
-        projectId: 'dummyEmptyProject',
-        httpService: fakeEmptyTestService,
-        globalQueryConfig: {
-          customHeaders: [
-            anotherHeader,
-          ],
-        },
-      };
-
-      await sourceNodes(
-        dummyCreation,
-        {
-          deliveryClientConfig,
-          languageCodenames: ['default'],
-        },
-      );
-
-      expect(deliveryClientConfig.globalQueryConfig.customHeaders)
-        .toContainEqual(customTrackingHeader);
-      expect(deliveryClientConfig.globalQueryConfig.customHeaders)
-        .toContainEqual(anotherHeader);
-      expect(deliveryClientConfig.globalQueryConfig.customHeaders.length)
-        .toEqual(2);
-    });
-  });
 
   describe('complex multilingual data section', () => {
     const fakeComplexConfig = new Map();
@@ -186,9 +56,11 @@ describe('sourceNodes', () => {
 
     const createNodeMock = jest.fn();
     const createTypesMock = jest.fn();
-    const mockedSchema = { buildObjectType: jest.fn((input) => ({
-      data: input,
-    }))};
+    const mockedSchema = {
+      buildObjectType: jest.fn((input) => ({
+        data: input,
+      })),
+    };
 
     const actions = {
       actions: {
@@ -203,7 +75,7 @@ describe('sourceNodes', () => {
       projectId: 'dummyProject',
       typeResolvers: [],
       httpService: new KontentTestHttpService(
-        fakeComplexConfig,
+        fakeComplexConfig
       ),
     };
 
@@ -220,5 +92,3 @@ describe('sourceNodes', () => {
     });
   });
 });
-
-
